+++ components/collector/src/source_collectors/api_source_collectors/azure_devops.py
@@ -142,6 +142,8 @@ class AzureDevopsTests(SourceCollector):
     async def _parse_source_responses(self, responses: SourceResponses) -> SourceMeasurement:
         test_results = cast(List[str], self._parameter("test_result"))
         test_run_names_to_include = cast(List[str], self._parameter("test_run_names_to_include")) or ["all"]
+        test_run_states_to_include = [
+            value.lower() for value in self._parameter("test_run_states_to_include")] or ["all"]
         runs = (await responses[0].json()).get("value", [])
         highest_build_nr_seen: Dict[str, int] = defaultdict(int)
         highest_build_nr_test_runs: Dict[str, Entities] = defaultdict(list)
@@ -150,6 +152,9 @@ class AzureDevopsTests(SourceCollector):
             name = run.get("name", "Unknown test run name")
             if test_run_names_to_include != ["all"] and \
                     not match_string_or_regular_expression(name, test_run_names_to_include):
+                continue
+            state = run.get("state", "Unknown test run state")
+            if test_run_states_to_include != ["all"] and state.lower() not in test_run_states_to_include:
                 continue
             build_nr = int(run.get("build", {}).get("id", "-1"))
             if build_nr < highest_build_nr_seen[name]:
@@ -161,9 +166,9 @@ class AzureDevopsTests(SourceCollector):
             highest_build_test_count[name] += sum(run.get(test_result, 0) for test_result in test_results)
             highest_build_nr_test_runs[name].append(
                 dict(
-                    key=run["id"], name=run.get("name", "Unknown test run name"), state=run.get("state", ""),
-                    build_id=str(build_nr), url=run.get("webAccessUrl", ""), started_date=run.get("startedDate", ""),
-                    completed_date=run.get("completedDate", ""), incomplete_tests=str(run.get("incompleteTests", 0)),
+                    key=run["id"], name=name, state=state, build_id=str(build_nr), url=run.get("webAccessUrl", ""),
+                    started_date=run.get("startedDate", ""), completed_date=run.get("completedDate", ""),
+                    incomplete_tests=str(run.get("incompleteTests", 0)),
                     not_applicable_tests=str(run.get("notApplicableTests", 0)),
                     passed_tests=str(run.get("passedTests", 0)), unanalyzed_tests=str(run.get("unanalyzedTests", 0)),
                     total_tests=str(run.get("totalTests", 0))))