+++ components/collector/tests/source_collectors/api_source_collectors/test_azure_devops.py
@@ -129,6 +129,7 @@ class AzureDevopsTestsTest(AzureDevopsTestCase):
         """Test that the number of tests is returned."""
         self.sources["source_id"]["test_result"] = []
         self.sources["source_id"]["parameters"]["test_run_names_to_include"] = ["A.*"]
+        self.sources["source_id"]["parameters"]["test_run_states_to_include"] = ["completed"]
         metric = dict(type="tests", sources=self.sources, addition="sum")
         response = await self.collect(
             metric, get_request_json_return_value=dict(
@@ -140,7 +141,8 @@ class AzureDevopsTestsTest(AzureDevopsTestCase):
                     dict(id="3", name="A", build=dict(id="1"), state="Completed", passedTests=4, totalTests=4),
                     dict(id="4", name="A", build=dict(id="2"), state="Completed", passedTests=1, totalTests=1),
                     dict(id="5", name="B", build=dict(id="3"), state="Completed", passedTests=5, totalTests=5),
-                    dict(id="6", name="A+", build=dict(id="1"), state="Completed", passedTests=6, totalTests=6)]))
+                    dict(id="6", name="A+", build=dict(id="1"), state="Completed", passedTests=6, totalTests=6),
+                    dict(id="7", name="A+", build=dict(id="2"), state="InProgress", passedTests=6, totalTests=6)]))
         self.assert_measurement(response, value="10", entities=[
             dict(key="2", name="A", state="Completed", build_id="2", started_date="2020-06-20T14:56:00.58Z",
                  completed_date="2020-06-20T14:56:00.633Z", incomplete_tests="0", not_applicable_tests="1",
@@ -157,7 +159,9 @@ class AzureDevopsTestsTest(AzureDevopsTestCase):
         self.sources["source_id"]["test_result"] = ["failed"]
         metric = dict(type="tests", sources=self.sources, addition="sum")
         response = await self.collect(
-            metric, get_request_json_return_value=dict(value=[dict(id="1", build=dict(id="1"), unanalyzedTests=4)]))
+            metric,
+            get_request_json_return_value=dict(
+                value=[dict(id="1", build=dict(id="1"), state="Completed", unanalyzedTests=4)]))
         self.assert_measurement(response, value="4")